frontend/requirements.txt
streamlit
requests
google-auth
google-auth-oauthlib
Pillow
